+++ tomcat/src/main/java/org/apache/coyote/http11/session/Session.java
@@ -23,12 +23,4 @@ public class Session {
     public void setAttribute(final String name, final Object value) {
         values.put(name, value);
     }
-
-    public void removeAttribute(final String name) {
-        values.remove(name);
-    }
-
-    public void invalidate() {
-        values.clear();
-    }
 }
